Update Category CRUD

+++ src/main/java/com/teamseven/ticketresell/repository/CategoryRepository.java
@@ -4,7 +4,7 @@ import com.teamseven.ticketresell.entity.CategoryEntity;
 import com.teamseven.ticketresell.entity.TicketEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CategoryRepository extends JpaRepository<CategoryEntity, Integer> {
+public interface CategoryRepository extends JpaRepository<CategoryEntity, Long> {
     CategoryEntity findById(long id);
 
 }
